+++ src/api/index.ts
@@ -69,5 +69,4 @@ export {
   adminClient,
   rpcClient,
   blobClient,
-  ApiClient,
 };
