+++ src/experimental/index.ts
@@ -1,6 +1,3 @@
-export {
-  CalimeroProvider,
-  useCalimero,
-} from './CalimeroProvider';
+export { CalimeroProvider, useCalimero } from './CalimeroProvider';
 export { default as CalimeroConnectButton } from './CalimeroConnectButton';
 export * from './types';
